Room/User Management
There is a "rooms" object in server. Each item of that object is going to be a room. Each room has a list of users, and the room owner.

When someone creates a room, the room is created and the user is added to the list of users. the roomowner is assigned to that user that created the room.

When someone joins a room, that user is added to the list of users of that particular room.

Right now, you have to refresh the room to then get the most updated info on users active