Live Status of Users in a Room
I installed the socket.io (module?) that will allow us to see real-time updates of our room details.

Now everyone in the room can see people coming and going without needing to refresh (theyre seeing it live).

This works in the following way. when someone clicks on join room from the landing page, the request handler in the server then sends a "broadcast" to all clients in that room with the data of the current user that just joined. the room.html then has a "request handler" for this "broadcast" which updates the current list with the new users data.

changed: app.listen(port) to server.listen(port).