Removing Users and Rooms upon Disconnection
Made it so when a user disconnects, theyre removed from the user list. Once the server detects that someone has disconnected, it gets what room they were in and what user it was, and then it sends a message to all people in that room so that each client can automatically update the user list of that room.

Also, if the room is empty, then the room object is removed and the room isnt active anymore.