Attempted to fix game fucntionality
the game doesn't go into the second round. I tried to use various different approached and included log statements to locate where the problem was but I could figure out how to proceed with this. I am pushing the latest version of each file I made changes to log progress but in reality this was just one of the weeks where I have not made significant progress, now I am planning to sit with the team to discuss how to tackle this problem and work in it tohgether