Choosing User-Created Decks
Now you can use user-created decks.

After the page loads, a function is ran to load up to 3 decks. there is also a load more button, which call this function when its pressed.

I changed the start-game post request to also consider whether or not the user chose to play with a user-created deck.

Choosing user-created decks has priority over creating your own.

Added another property to gamestate that determines whether or  not youre using user created decks, so as to not create a new one.