Added Admin Tool To Delete Deck
Just for debugging, added a tool that can help us delete decks that we're not actually using, or that we used for testing to see if the "create custom deck" system works fine.

A button is generated as every custom deck is displayed, with the value of that deck id. press the right combination to show them, and if you press them, a request is sent to the server to delete a deck with that particular deck id from the custom deck table on the database.