Only Room Creator Can Start Game
I made it so only the room creator can start the game.

Every room fetches the room creator and compares it against the username of the current room.

I also had to fix an issue where I had userId meaning the code on the server, but in the room.html, userId wsa for the username. I made it consistent and readable so now username always means username (user picked) and the userId always means the code thats randomly generated for every user.

Also, cleaned out some unnecessary debug statements