Fixed Join Room Error Alerts and Room Layour
Fixed a bug where the user wasn't seeing the alerts when they joined a room and there was an error (room is full error and room not found error)

Also rearranged a little bit the layour of room.html. And made sure that the set max number of players is only visible only for the room owner.